fix(dcd): resolve around **/AddTokenPanel
