+++ src/components/[guild]/RolePlatforms/components/AddRoleRewardModal/components/AddTokenPanel/AddTokenPanel.tsx
@@ -26,34 +26,6 @@ import PoolStep from "./components/PoolStep"
 import SetTokenStep from "./components/SetTokenStep"
 import TokenAmountStep from "./components/TokenAmountStep"
 
-export enum TokenRewardType {
-  DYNAMIC_SNAPSHOT = "DYNAMIC_SNAPSHOT",
-  DYNAMIC_POINTS = "DYNAMIC_POINTS",
-  STATIC = "STATIC",
-}
-
-export type AddTokenFormType = {
-  poolId: number
-  multiplier: number
-  addition: number
-  chain: Chain
-  contractAddress: `0x${string}`
-  tokenAddress: `0x${string}`
-  name: string
-  description: string
-  imageUrl: string
-  data: {
-    guildPlatformId: number
-  }
-  snapshotId: number
-  type: TokenRewardType
-  staticValue?: number
-  snapshotRequirement?: Extract<
-    Schemas["RequirementCreationPayload"],
-    { type: "GUILD_SNAPSHOT" }
-  >
-}
-
 const AddTokenPanel = ({ onAdd }: AddRewardPanelProps) => {
   const methods = useForm<AddTokenFormType>({
     mode: "all",
